+++ dataforge-data/src/jvmMain/kotlin/space/kscience/dataforge/data/dataFilterJvm.kt
@@ -6,6 +6,8 @@ import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.names.matches
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
 import kotlin.reflect.KType
 import kotlin.reflect.full.isSubtypeOf
 import kotlin.reflect.typeOf
@@ -28,46 +30,47 @@ private fun <R : Any> Data<*>.castOrNull(type: KType): Data<R>? =
  * Select all data matching given type and filters. Does not modify paths
  *
  * @param namePattern a name match patter according to [Name.matches]
- * @param filter addition filtering condition based on item name and meta. By default, accepts all
+ * @param predicate addition filtering condition based on item name and meta. By default, accepts all
  */
 @OptIn(DFExperimental::class)
-public fun <R : Any> DataSet<*>.select(
+public fun <R : Any> DataSet<*>.filterIsInstance(
     type: KType,
-    namePattern: Name? = null,
-    filter: (name: Name, meta: Meta) -> Boolean = { _, _ -> true },
-): ActiveDataSet<R> = object : ActiveDataSet<R> {
+    predicate: (name: Name, meta: Meta) -> Boolean = { _, _ -> true },
+): DataSource<R> = object : DataSource<R> {
     override val dataType = type
 
-    override val meta: Meta get() = this@select.meta
+    override val coroutineContext: CoroutineContext
+        get() = (this@filterIsInstance as? DataSource)?.coroutineContext ?: EmptyCoroutineContext
+
+    override val meta: Meta get() = this@filterIsInstance.meta
 
     private fun checkDatum(name: Name, datum: Data<*>): Boolean = datum.type.isSubtypeOf(type)
-            && (namePattern == null || name.matches(namePattern))
-            && filter(name, datum.meta)
+            && predicate(name, datum.meta)
 
-    override fun dataSequence(): Sequence<NamedData<R>> = this@select.dataSequence().filter {
+    override fun dataSequence(): Sequence<NamedData<R>> = this@filterIsInstance.dataSequence().filter {
         checkDatum(it.name, it.data)
     }.map {
         @Suppress("UNCHECKED_CAST")
         it as NamedData<R>
     }
 
-    override fun get(name: Name): Data<R>? = this@select[name]?.let { datum ->
+    override fun get(name: Name): Data<R>? = this@filterIsInstance[name]?.let { datum ->
         if (checkDatum(name, datum)) datum.castOrNull(type) else null
     }
 
-    override val updates: Flow<Name> = this@select.updates.filter {
-        val datum = this@select[it] ?: return@filter false
-        checkDatum(it, datum)
+    override val updates: Flow<Name> = this@filterIsInstance.updates.filter { name ->
+        get(name)?.let { datum ->
+            checkDatum(name, datum)
+        } ?: false
     }
 }
 
 /**
  * Select a single datum of the appropriate type
  */
-public inline fun <reified R : Any> DataSet<*>.select(
-    namePattern: Name? = null,
-    noinline filter: (name: Name, meta: Meta) -> Boolean = { _, _ -> true },
-): DataSet<R> = select(typeOf<R>(), namePattern, filter)
+public inline fun <reified R : Any> DataSet<*>.filterIsInstance(
+    noinline predicate: (name: Name, meta: Meta) -> Boolean = { _, _ -> true },
+): DataSet<R> = filterIsInstance(typeOf<R>(), predicate)
 
 /**
  * Select a single datum if it is present and of given [type]