+++ dataforge-workspace/src/jvmMain/kotlin/space/kscience/dataforge/workspace/workspaceJvm.kt
@@ -1,21 +1,24 @@
 package space.kscience.dataforge.workspace
 
-import kotlinx.coroutines.runBlocking
 import space.kscience.dataforge.data.DataSet
-import space.kscience.dataforge.data.DataSetBuilder
-import space.kscience.dataforge.data.select
+import space.kscience.dataforge.data.filterIsInstance
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.names.Name
+import space.kscience.dataforge.names.matches
 
-public fun WorkspaceBuilder.data(builder: suspend DataSetBuilder<Any>.() -> Unit): Unit = runBlocking {
-    buildData(builder)
-}
+//public fun WorkspaceBuilder.data(builder: DataSetBuilder<Any>.() -> Unit): Unit = runBlocking {
+//    data(builder)
+//}
 
-public inline fun <reified T: Any> TaskResultBuilder<*>.data(namePattern: Name? = null): DataSelector<T> = object : DataSelector<T> {
-    override suspend fun select(workspace: Workspace, meta: Meta): DataSet<T> = workspace.data.select(namePattern)
-}
+public inline fun <reified T : Any> TaskResultBuilder<*>.data(namePattern: Name? = null): DataSelector<T> =
+    object : DataSelector<T> {
+        override suspend fun select(workspace: Workspace, meta: Meta): DataSet<T> =
+            workspace.data.filterIsInstance { name, _ ->
+                namePattern == null || name.matches(namePattern)
+            }
+    }
 
 public suspend inline fun <reified T : Any> TaskResultBuilder<*>.fromTask(
     task: Name,
     taskMeta: Meta = Meta.EMPTY,
-): DataSet<T> = workspace.produce(task, taskMeta).select()
+): DataSet<T> = workspace.produce(task, taskMeta).filterIsInstance()